PRT_updater2/requirements.txt
pyodbc==4.0.39
pandas==2.1.0
openpyxl==3.1.2
mysqlclient==2.2.4
mysql-connector-python==9.0.0
requests==2.32.3
python-dotenv==1.0.1